Downstream updates following rename of IconMatcher to IconDetector
