+++ sto_sister/stages/__init__.py
@@ -4,7 +4,7 @@ from .icon_group_locator import IconGroupLocatorStage
 from .icon_slot_locator import IconSlotLocatorStage
 from .icon_prefilter import IconPrefilterStage
 from .icon_overlay_detector import IconOverlayDetectorStage
-from .icon_matching import IconMatchingStage
+from .icon_detector import IconDetectorStage
 from .output_transform import OutputTransformationStage
 
 __all__ = [
@@ -14,6 +14,6 @@ __all__ = [
     "IconSlotLocatorStage",
     "IconPrefilterStage",
     "IconOverlayDetectorStage",
-    "IconMatchingStage",
+    "IconDetectorStage",
     "OutputTransformationStage",
 ]
