+++ sto_sister/stages/icon_detector.py
